Added partial support for Suse to fail2ban role.
On RedHat-likes, a more or less complete and working install is delivered.
On Suse, only the basic fail2ban install is put into place.
This is in part so I can check into what firewalling options are available for Suse-likes.
I personally don't make use of SuSEFirewall, in part because Suse can run other, more generic, iptables management scripts.
tl;dr: fail2ban is present, but not doing much of any use on Suse, and will be made to do more soon.

Signed-off-by: I was eaten by a grue <iwaseatenbyagrue@gmail.com>