added role to install and configure consul.
By default, a client agent is put in place, that seeks to join any nodes found within the consul_server group.
The domain, datacenter, and bind/advertise address are configurable via variables.

When a node is part of the consul_server group, consul's webui is also installed, and bootstrap settings are configured.
However, clustering consul servers isn't yet tested, and some config switches are lacking for it to fully work - which will be addressed soon.

Signed-off-by: I was eaten by a grue <iwaseatenbyagrue@gmail.com>