Added a link between the downloading of a new release, and extracting the relevant binaries, for all components in prometheus. Also added restart handlers for all prometheus components.
Downloading a new release will now force a new symlink to be created for all the prometheus components.
Briefly, this works by registering the result of the download step.
Subsequent steps that (for prometheus) create links to the relevant binaries will check that result, and if a change is detected, will force the creation of a new symlink.
This is not yet tested in the context of the prometheus role, but works in similar situations for the consul role.

Also created a specific restart handler for the alertmanagement, node_exporter, and prometheus_server.

Signed-off-by: I was eaten by a grue <iwaseatenbyagrue@gmail.com>