Rewrote some parts of the prometheus role to simplify it, and try and make it more usable.
This notably involved changing from checking group memberships directly to determine which components are needed, in favour of using switches instead.
In practice, a given task might previously have a when clause something like this: 'thing' in group_names.
This now uses a switch instead, e.g. when: not skip_thing.
The switches for prometheus are still based on groups (which have been renamed to make them plural), but can be overriden by variables.
The groups are still used for auto-configuring some aspects of the setup (and the prometheus firewall rules will soon do so), but you can now potentially configure a standby node by toggling some switches, and without having the standby node used by anything (because it wouldn't be part of e.g. prometheus_servers).

Signed-off-by: I was eaten by a grue <iwaseatenbyagrue@gmail.com>