Added ipset-based rule for grafana to speak to prometheus servers (in prometheus role).
The default deployment for prometheus server and grafana collocates them (because I happen to prefer grafana to prometheus's web front-end).
However, for cases  where they are not collocated, and also to allow grafana to use a cluster of prometheus servers if wanted, the firewalld tasks will allow the required access.

Signed-off-by: I was eaten by a grue <iwaseatenbyagrue@gmail.com>